Feat: Letter Controller 기능 추가

+++ Lemon/backend/src/main/java/com/lemon/backend/domain/letter/repository/LetterRepository.java
@@ -4,5 +4,5 @@ import com.lemon.backend.domain.letter.entity.Letter;
 import com.lemon.backend.domain.letter.repository.custom.LetterRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LetterRepository extends JpaRepository<Letter, Integer>, LetterRepositoryCustom {
+public interface LetterRepository extends JpaRepository<Letter, Long>, LetterRepositoryCustom {
 }
